coap_client: fixes for synchronous instability
Previously, the way semaphores were used in the synchronous
client transactions was incorrect, causing undefined behavior
in cases where the semaphore was accessed after the *_sync API
had already returned. This caused spurious crashes and test
failures.

To resolve the issue, the *_sync APIs now use a FreeRTOS
event group, for the CoAP task to notify the user task that
the request has been completed, with one event group bit
to indicate "response received" and another for "timeout".
The ownership of the event group is transferred from user task
to CoAP task at the time of enqueueing the request, and the
CoAP task frees the event group after the request is processed.

To safely delete the event group, the CoAP task must know
when the user task has received the event group notification,
so an additional semaphore is added for the user task to
notify the CoAP thread when the event has been received, and
it's safe to delete both the event group and semaphore.

+++ components/golioth_sdk/golioth_ota.c
@@ -32,7 +32,7 @@ const golioth_ota_component_t* golioth_ota_find_component(
         const char* package) {
     // Scan the manifest until we find the component with matching package.
     const golioth_ota_component_t* found = NULL;
-    for (int i = 0; i < manifest->num_components; i++) {
+    for (size_t i = 0; i < manifest->num_components; i++) {
         const golioth_ota_component_t* c = &manifest->components[i];
         bool matches = (0 == strcmp(c->package, package));
         if (matches) {
